Add validation to file upload
